requirements-docker.txt
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dicom2nifti==2.6.1
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.2
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.6.2
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.1
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
nilearn==0.11.1
numpy==1.26.4
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydicom==3.0.1
pyparsing==3.2.2
python-dateutil==2.9.0.post0
python-gdcm==3.0.24.1
pytz==2025.2
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
SimpleITK==2.4.1
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
